+++ src/app/sf-form-wrapper/sf-form-wrapper.component.ts
@@ -9,7 +9,13 @@ import {
 } from '@angular/core';
 import {FormService} from '../services/form.service';
 import {LinkIdCollection} from '../item/item.component';
-import {ArrayProperty, FormComponent, FormProperty, PropertyGroup, ObjectProperty} from '@lhncbc/ngx-schema-form';
+import {
+  ArrayProperty,
+  FormComponent,
+  FormProperty,
+  PropertyGroup,
+  ObjectProperty,
+} from '@lhncbc/ngx-schema-form';
 import {ExtensionsService} from '../services/extensions.service';
 import {Util} from '../lib/util';
 import { SharedObjectService } from '../services/shared-object.service';
@@ -169,9 +175,14 @@ export class SfFormWrapperComponent implements OnInit, OnChanges, AfterViewInit 
     const questionItem = this.formService.getTreeNodeByLinkId(q.value);
 
     let aType = '';
+    let invalid = true;
 
     if (questionItem) {
       aType = questionItem.data.type;
+      const validSources = this.formService.getSourcesExcludingFocusedTree();
+      invalid = !validSources.some((source) => {
+        return (source.data.linkId === questionItem.data.linkId);
+      });
     }
     // The condition key is used to differentiate between each enableWHen conditions.
     let condKey = '';
@@ -212,6 +223,7 @@ export class SfFormWrapperComponent implements OnInit, OnChanges, AfterViewInit 
       'q': q,
       'qItem': this.validationService.populateQuestionItem(questionItem),
       'aType': aType,
+      'invalid': invalid,
       'answerTypeProperty': answerType,
       'op': op,
       'aField': aField,
@@ -270,7 +282,7 @@ export class SfFormWrapperComponent implements OnInit, OnChanges, AfterViewInit 
    *          3. (ENABLEWHEN_ANSWER_REQUIRED)  - The question is provided and valid, the operator is provided and not
    *                                             and not equal to 'exists', and the answer is empty.
    */
-  validateEnableWhenAll (value: any, arrayProperty: ArrayProperty, rootProperty: PropertyGroup): any[] | null {
+  validateEnableWhenAll(value: any, arrayProperty: ArrayProperty, rootProperty: PropertyGroup): any[] | null {
     let errors = null;
     // iterate all properties
     arrayProperty.forEachChild((property: ObjectProperty) => {
@@ -297,7 +309,7 @@ export class SfFormWrapperComponent implements OnInit, OnChanges, AfterViewInit 
    *          3. (ENABLEWHEN_ANSWER_REQUIRED)  - The question is provided and valid, the operator is provided and not
    *                                             and not equal to 'exists', and the answer is empty.
    */
-  validateEnableWhenSingle (value: any, formProperty: ObjectProperty, rootProperty: PropertyGroup): any[] | null {
+  validateEnableWhenSingle(value: any, formProperty: ObjectProperty, rootProperty: PropertyGroup): any[] | null {
     let errors: any[] = [];
 
     if (!this.model) {